moved the mapper from mapper to controller
